+++ NEWS.md
@@ -16,6 +16,8 @@ New features(Analysis):
 + Infer from `return new static();` and `return $this;` that the return type of a method is `@return static`, not `@return self` (#2797)
   (and propogate that to inherited methods)
 + Fix some false positives when casting array types containing `static` to types containing the class or its ancestors. (#2797)
++ Add `PhanTypeInstantiateAbstractStatic` and `PhanTypeInstantiateTraitStaticOrSelf` as lower-severity warnings about `return new self()` and `return new static()` (#2797)
+  (emitted in static methods of abstract classes)
 
 Language Server/Daemon mode:
 + Add `--language-server-min-diagnostics-delay-ms <ms>`, to work around race conditions in some language clients.